pyproject.toml
[tool.poetry]
name = "weatherapi"
version = "0.1.0"
description = ""
authors = ["simone <134292096+Wellington9959@users.noreply.github.com>"]
license = "GNU GENERAL PUBLIC LICENSE v3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
redis = "^5.2.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
flask-limiter = {extras = ["redis"], version = "^3.9.2"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.292"
black = "^23.9.1"

[tool.black]
line-length = 88

[tool.ruff]
lint.select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
lint.fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401", # unused import
    "F403", # star imports
]
"base.py" = [
    "F401", # unused import
    "F403", # star imports
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
